Finishing part4

+++ part4/bloglist/controllers/blog.js
@@ -1,8 +1,13 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
   })
 
@@ -14,24 +19,46 @@ blogsRouter.get('/:id', async (request, response) => {
     response.status(404).end()
   }
 }) 
+
   
 blogsRouter.post('/', async (request, response) => {
+  const user = request.user
   const body = request.body
+  const token = request.token
+
+  if(!token || !user.id) {
+    return response.status(401).json({ error: 'token missing or invalid' }) 
+  }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user._id,
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if(!request.token || !user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  if(blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'token missing or invalid' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
